requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic
python-dotenv==1.0.1
loguru==0.7.2

pandas
numpy
scikit-learn
xgboost
shap
matplotlib

google-cloud-bigquery==3.25.0
google-auth==2.35.0

langchain==0.3.3
langchain-core==0.3.10
langchain-community==0.3.2

streamlit==1.39.0
requests==2.32.3

# Frontend is separate (npm). This file is for Python deps only.